+++ src/lib.rs
@@ -133,6 +133,7 @@ pub(crate) use debug_assert_bounds;
 
 /// Initializes the `env_logger` logger with a custom format including
 /// timestamps with elapsed time since initialization.
+#[cfg(feature = "cli")]
 pub fn init_env_logger() -> anyhow::Result<()> {
     use jiff::{
         SpanRound,